adds a hotkey to launch the [neovide](https://github.com/tongson/neovide) application from the hammerspoon config. when the hotkey is pressed, neovide will launch or be brought to the front. if neovide is already open but hidden, it will be brought to the front.
the hotkey is `ctrl+shift+z`, and it is defined in the `init.lua` file. the hotkey will check if neovide is open, and if not it will launch it, or if it is open it will bring it to the front. if the main window of neovide is not in the current space, it will be moved to the current space before being brought to the front.